Branch commit
Refine the  functions and folder structure
